move RunModule to seperate file
